Initial code
There is no method to get draft releases, so by using this code we can
create a draft release earlier in the pipeline and when we are ready we
can convert it to a published release.

This code is used to find the draft release using the API. You need to
pass in a GITHUB_TOKEN with access to the repository because draft
releases are usually kept private.