Users to Avoid
When a PR is created by Snyk or Dependabot quite a lot of text is
provided, that our community are not interested in and it floods our
slack channels. They are happy to see the title of the PR instead.
This change allows a list of users to be passed in, which if the PR is
created by one of those users the title is chosen rather than the body.
The defaults Snyk and Dependabot